Add bounds checking for decoding CDMA SMS using Shift-JIS.
The test case testRandomUserData() in CdmaSmsCbTest fails since we
added support for decoding Shift-JIS messages. Change decodeShiftJis()
to call the helper method decodeCharset(), which checks the array
bounds and tries to decode as much as possible if the length extends
past the end of the PDU, instead of throwing
StringIndexOutOfBoundsException.

Bug: 8107987
Change-Id: I180e8360e391cd8276752a44a58a9c3a635dc284