Add SMS filtering intent and API for message deduping by the carrier app.
If a carrier app is active, the SMS_FILTER intent is sent for every
incoming SMS message. If the carrier app considers this message as a
duplicate, it may set a status code to filter out this SMS. It may
also set a pdu in result extras to overwrite the pdu.

Change-Id: I41526b21df31142b8f6373f3aefbe360c5938af3